Fixed chunk upload
We broke it when we switched from ResumableJS to FlowJS. Oops. I
think what happened is I tested it on what I thought was a big
file but it was really a small one so none of it got chunked out.

The effect I was seeing was that we'd be getting a new file for
each chunk instead of one big file at the end.