Changed how we do Host auth
Now we are using a personal access token for both users and hosts. We
can safely do this now because the identifiers are UUIDs. Previously,
if we had a route that was marked "auth:api,host" then if a host token
for id 1 came along, it could access the user 1. This has a lot to do
with the internals of Laravel Passport. Now, they'll have separate UUIDs
so the user lookup for the api won't work, but it will match for the host
auth.

This is super useful for things like the broadcasting auth route
because we can mark it as "auth:web,api,host" and it will try to auth
in that order and won't have a security risk.

It also means we can remove a ton of our clunky host grants, extra client
types, HostManager, etc. This _really_ simplifies so much that I'm going
to be happy about it for weeks.